pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autoharness"
version = "0.1.0"
description = "Autonomous harness optimizer for agent evaluation campaigns"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
autoharness = "autoharness.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/autoharness"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
    "-q",
]
